+++ test/libsolidity/syntaxTests/types/call_unimplemented_internal_function.sol
@@ -1,0 +1,6 @@
+abstract contract A {
+    function f() public virtual;
+    function g() public {
+        f();
+    }
+}