+++ test/libsolidity/syntaxTests/types/contractTypeType/members/base_contract.sol
@@ -1,0 +1,11 @@
+contract B {
+    function f() external {}
+    function g() public {}
+}
+contract C is B {
+    function h() public {
+        B.f.selector;
+        B.g.selector;
+        B.g();
+    }
+}