+++ test/libsolidity/syntaxTests/types/contractTypeType/members/call_unimplemented_base.sol
@@ -1,0 +1,10 @@
+abstract contract B {
+    function f() public virtual;
+}
+contract C is B {
+    function f() public override {
+        B.f();
+    }
+}
+// ----
+// TypeError: (118-123): Cannot call unimplemented base function.