+++ test/libsolidity/syntaxTests/types/library_internal_call.sol
@@ -1,0 +1,4 @@
+library L {
+    function a() public pure {}
+    function b() public pure { a(); }
+}